+++ CHANGELOG.md
@@ -29,6 +29,7 @@
 * Add Delete and Update to ES sink [#822](https://github.com/tremor-rs/tremor-runtime/issues/822)
 * Add more metadata to Kafka source [#874](https://github.com/tremor-rs/tremor-runtime/issues/874)
 * Update to simd-json 0.4
+* Add tests covering basic operations and string interpolation for tremor-script[#721](https://github.com/tremor-rs/tremor-runtime/issues/721) 
 
 ### Fixes